+++ config/fmcsa.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'api_key' => env('FMCSA_API_KEY'),
+];
