+++ resources/js/Pages/Carriers/Index.tsx
@@ -14,32 +14,11 @@ import { Carrier } from '@/types';
 import { Head, router } from '@inertiajs/react';
 import { Plus } from 'lucide-react';
 import { useState } from 'react';
+import CarrierCreateDialog from './Partials/CarrierCreateDialog';
 
 export default function Index() {
     const [isOpen, setIsOpen] = useState(false);
-    const [carrierName, setCarrierName] = useState('');
-    const [errors, setErrors] = useState<Record<string, string>>({});
 
-    const handleSubmit = (e: React.FormEvent) => {
-        e.preventDefault();
-        setErrors({});
-        router.post(
-            route('carriers.store'),
-            {
-                name: carrierName,
-            },
-            {
-                preserveScroll: true,
-                onSuccess: () => {
-                    setIsOpen(false);
-                    setCarrierName('');
-                },
-                onError: (errors) => {
-                    setErrors(errors);
-                },
-            },
-        );
-    };
 
     const setSelectedCarrier = (carrier: Carrier) => {
         router.visit(route('carriers.show', { carrier }));
@@ -60,7 +39,6 @@ export default function Index() {
                         type="button"
                         onClick={() => {
                             setIsOpen(true);
-                            setCarrierName('');
                         }}
                     >
                         <Plus /> Create Carrier
@@ -70,31 +48,7 @@ export default function Index() {
                     <CarrierList onSelect={setSelectedCarrier} />
                 </div>
             </div>
-            <Dialog open={isOpen} onOpenChange={setIsOpen}>
-                <DialogContent>
-                    <DialogHeader>
-                        <DialogTitle>Create Carrier</DialogTitle>
-                    </DialogHeader>
-                    <form onSubmit={handleSubmit} className="space-y-4">
-                        <Input
-                            placeholder="Carrier name"
-                            value={carrierName}
-                            onChange={(e) => setCarrierName(e.target.value)}
-                        />
-                        {errors.name && <InputError message={errors.name} />}
-                        <DialogFooter>
-                            <Button
-                                type="button"
-                                variant="outline"
-                                onClick={() => setIsOpen(false)}
-                            >
-                                Cancel
-                            </Button>
-                            <Button type="submit">Create</Button>
-                        </DialogFooter>
-                    </form>
-                </DialogContent>
-            </Dialog>
+            <CarrierCreateDialog isOpen={isOpen} setIsOpen={setIsOpen} />
         </AuthenticatedLayout>
     );
 }
